split messaging code
