+++ app/messaging/index.js
@@ -1,32 +1,8 @@
-const db = require('../data')
 const { MessageReceiver } = require('ffc-messaging')
-const NotifyClient = require('notifications-node-client').NotifyClient
 const { uuid } = require('uuidv4')
-
-const saveToDb = async (message) => {
-  try {
-    await db.messages.create({
-      scheme: message.body.scheme,
-      tags: message.body.tags,
-      crn: message.body.crn,
-      content: message.body.content,
-      requestedDate: message.body.requestedDate
-    })
-  } catch (error) {
-    console.error('ERROR: ', error)
-  }
-}
-
-const sendViaNotify = async (message) => {
-  const notifyClient = new NotifyClient(process.env.NOTIFY_API_KEY)
-  await notifyClient.sendEmail(process.env.NOTIFY_TEMPLATE_ID, 'rana.salem@defra.gov.uk', {
-    personalisation: {
-      heading: message.body.content.heading,
-      content: message.body.content.body
-    },
-    reference: uuid()
-  })
-}
+const db = require('../data')
+const { saveToDatabase } = require('./save-to-database')
+const { sendMessage } = require('./send-message')
 
 const handleMessage = async (message) => {
   console.log('Received message: ', message.body)
